+++ apps/gateway/src/lib/hooks.ts
